+++ PizzaTowerEscapeMusic/PluginInfo.cs
@@ -6,6 +6,6 @@
 
         public const string PLUGIN_NAME = "PizzaTowerEscapeMusic";
 
-        public const string PLUGIN_VERSION = "2.5.0";
+        public const string PLUGIN_VERSION = "2.5.1";
     }
 }