+++ CHANGELOG.md
@@ -1,10 +1,13 @@
+### 2.5.1
+- I'm must be too tired, fix typo in 2.5.0 changelog, and fix `ResetCounters` completely not working
+
 ### 2.5.0
-- Breaking changes to `Timer` condition, now timer won't reset by itself, it must be reset by `ResetTimer` script event
+- Breaking changes to `Timer` condition, now timer won't reset by itself, it must be reset by `ResetTimers` script event
 - Fixed GameEventType `FiringPlayers`, `GameOver` completly not working
 - Added options in VolumeModifier `volumeLerpSpeedScale` and `stoppingVolumeLerpSpeedScale`
 - Added option in PlayMusic `introMusicNames`, yes just like PAYDAY
 - Added conditionType `PlayerFearLevel`, `Counter`
-- Added scriptEventType `ResetCounter`
+- Added scriptEventType `ResetCounters`
 - Added config option `SelectLabelManually`, support LethalConfig
 
 ### 2.4.8